Created message validators for NewRound and RoundChange
