+++ consensus/ibft/src/main/java/tech/pegasys/pantheon/consensus/ibft/statemachine/RoundChangeManager.java
@@ -108,7 +108,7 @@ public class RoundChangeManager {
   }
 
   private boolean isMessageValid(final RoundChange msg) {
-    return roundChangeMessageValidator.validateMessage(msg.getSignedPayload());
+    return roundChangeMessageValidator.validateRoundChange(msg);
   }
 
   private RoundChangeStatus storeRoundChangeMessage(final RoundChange msg) {
