+++ consensus/ibft/src/main/java/tech/pegasys/pantheon/consensus/ibft/validation/RoundChangeMessageValidator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 ConsenSys AG.
+ * Copyright 2019 ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -12,124 +12,18 @@
  */
 package tech.pegasys.pantheon.consensus.ibft.validation;
 
-import tech.pegasys.pantheon.consensus.ibft.ConsensusRoundIdentifier;
-import tech.pegasys.pantheon.consensus.ibft.payload.PreparePayload;
-import tech.pegasys.pantheon.consensus.ibft.payload.PreparedCertificate;
-import tech.pegasys.pantheon.consensus.ibft.payload.ProposalPayload;
-import tech.pegasys.pantheon.consensus.ibft.payload.RoundChangePayload;
-import tech.pegasys.pantheon.consensus.ibft.payload.SignedData;
-import tech.pegasys.pantheon.ethereum.core.Address;
-
-import java.util.Collection;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import tech.pegasys.pantheon.consensus.ibft.messagewrappers.RoundChange;
 
 public class RoundChangeMessageValidator {
 
-  private static final Logger LOG = LogManager.getLogger();
-
-  private final MessageValidatorForHeightFactory messageValidatorFactory;
-  private final Collection<Address> validators;
-  private final long minimumPrepareMessages;
-  private final long chainHeight;
+  private final RoundChangePayloadValidator roundChangePayloadValidator;
 
   public RoundChangeMessageValidator(
-      final MessageValidatorForHeightFactory messageValidatorFactory,
-      final Collection<Address> validators,
-      final long minimumPrepareMessages,
-      final long chainHeight) {
-    this.messageValidatorFactory = messageValidatorFactory;
-    this.validators = validators;
-    this.minimumPrepareMessages = minimumPrepareMessages;
-    this.chainHeight = chainHeight;
+      final RoundChangePayloadValidator roundChangePayloadValidator) {
+    this.roundChangePayloadValidator = roundChangePayloadValidator;
   }
 
-  public boolean validateMessage(final SignedData<RoundChangePayload> msg) {
-
-    if (!validators.contains(msg.getAuthor())) {
-      LOG.info(
-          "Invalid RoundChange message, was not transmitted by a validator for the associated"
-              + " round.");
-      return false;
-    }
-
-    final ConsensusRoundIdentifier targetRound = msg.getPayload().getRoundIdentifier();
-
-    if (targetRound.getSequenceNumber() != chainHeight) {
-      LOG.info("Invalid RoundChange message, not valid for local chain height.");
-      return false;
-    }
-
-    if (msg.getPayload().getPreparedCertificate().isPresent()) {
-      final PreparedCertificate certificate = msg.getPayload().getPreparedCertificate().get();
-
-      return validatePrepareCertificate(certificate, targetRound);
-    }
-
-    return true;
-  }
-
-  private boolean validatePrepareCertificate(
-      final PreparedCertificate certificate, final ConsensusRoundIdentifier roundChangeTarget) {
-    final SignedData<ProposalPayload> proposalMessage = certificate.getProposalPayload();
-
-    final ConsensusRoundIdentifier proposalRoundIdentifier =
-        proposalMessage.getPayload().getRoundIdentifier();
-
-    if (!validatePreparedCertificateRound(proposalRoundIdentifier, roundChangeTarget)) {
-      return false;
-    }
-
-    final SignedDataValidator signedDataValidator =
-        messageValidatorFactory.createAt(proposalRoundIdentifier);
-    return validateConsistencyOfPrepareCertificateMessages(certificate, signedDataValidator);
-  }
-
-  private boolean validateConsistencyOfPrepareCertificateMessages(
-      final PreparedCertificate certificate, final SignedDataValidator signedDataValidator) {
-
-    if (!signedDataValidator.addSignedProposalPayload(certificate.getProposalPayload())) {
-      LOG.info("Invalid RoundChange message, embedded Proposal message failed validation.");
-      return false;
-    }
-
-    if (certificate.getPreparePayloads().size() < minimumPrepareMessages) {
-      LOG.info(
-          "Invalid RoundChange message, insufficient Prepare messages exist to justify "
-              + "prepare certificate.");
-      return false;
-    }
-
-    for (final SignedData<PreparePayload> prepareMsg : certificate.getPreparePayloads()) {
-      if (!signedDataValidator.validatePrepareMessage(prepareMsg)) {
-        LOG.info("Invalid RoundChange message, embedded Prepare message failed validation.");
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  private boolean validatePreparedCertificateRound(
-      final ConsensusRoundIdentifier prepareCertRound,
-      final ConsensusRoundIdentifier roundChangeTarget) {
-
-    if (prepareCertRound.getSequenceNumber() != roundChangeTarget.getSequenceNumber()) {
-      LOG.info("Invalid RoundChange message, PreparedCertificate is not for local chain height.");
-      return false;
-    }
-
-    if (prepareCertRound.getRoundNumber() >= roundChangeTarget.getRoundNumber()) {
-      LOG.info(
-          "Invalid RoundChange message, PreparedCertificate not older than RoundChange target.");
-      return false;
-    }
-    return true;
-  }
-
-  @FunctionalInterface
-  public interface MessageValidatorForHeightFactory {
-    SignedDataValidator createAt(final ConsensusRoundIdentifier roundIdentifier);
+  public boolean validateRoundChange(final RoundChange msg) {
+    return roundChangePayloadValidator.validateRoundChange(msg.getSignedPayload());
   }
 }
